requirements_dev.txt
-r requirements.txt
black
flake8
mypy
pre-commit
pytest
pytest-homeassistant-custom-component
pylint-pytest
pylint
pytest-aiohttp
pytest-asyncio
types-aiohttp
